modulo abreviados terminado

+++ app/Model/Abbreviate.php
@@ -1,5 +1,6 @@
 <?php
 App::uses('AppModel', 'Model');
+App::uses('File', 'Utility');
 /**
  * Abbreviate Model
  *
@@ -13,6 +14,8 @@ class Abbreviate extends AppModel {
 	public $displayField = 'abbreviate';
 
 	var $path = 'files/numeros_abreviados.txt';
+	
+	var $updateFromfile = false;
 	
 	public $validate = array(
 		'abbreviate' => array(
@@ -44,18 +47,30 @@ class Abbreviate extends AppModel {
 			),
 		),
 	);
-
+	
 	public function fillDBFromFile() {
-		$this -> query('TRUNCATE TABLE abbreviates');
-		$data = $this -> readFile();
-		return $this -> saveAll($data);
+		$this->updateFromfile = false;
+		if($data = $this->readFile()) {
+			$this->updateFromfile=true;
+			$this->query('TRUNCATE TABLE abbreviates');
+			$success = true;
+			foreach ($data as $key => $value) {
+				if(!$this -> save($value)) $success = false;
+			}
+			$this->updateFromfile=false;
+			return $success;
+		} else {
+			return false;
+		}
 	}
 
-	private function readFile() {
+	public function readFile() {
 		$file = new File($this -> path, false, 0644);
-		$data = false;
-		if ($file -> exists()) {
+		$data = null;
+		if ($file -> exists() && $file -> readable()) {
 			$data = $file -> read();
+		} else {
+			return false;
 		}
 		$file -> close();
 		if ($data) {
@@ -64,11 +79,41 @@ class Abbreviate extends AppModel {
 			foreach ($tmp_data as $id => $abbreviate_info) {
 				if (!empty($abbreviate_info)) {
 					$info = explode('|', $abbreviate_info);
-					$data[] = array('Abbreviate' => array('id' => $id, 'abbreviate' => '*' . $info[0], 'phone' => $info[1]));
+					$data[] = array('Abbreviate' => array('id' => $id, 'abbreviate' => '*' . trim($info[0]), 'phone' => trim($info[1])));
 				}
 			}
 		}
 		return $data;
+	}
+
+	private function writeFile() {
+		$file = new File($this -> path, true, 0644);
+		if ($file -> exists() && $file->delete()) {
+			$file->create();
+			$abreviados = $this -> find('all');
+			foreach ($abreviados as $key => $value) {
+				$texto = $value['Abbreviate']['abbreviate'] . '|' . $value['Abbreviate']['phone'].chr(13);
+				$file -> append($texto);
+			}
+		} else {
+			// TODO : Hacer algo si no existe el archivo? (¡antes de crearlo!)
+		}
+		$file -> close();
+	}
+	
+	public function afterSave($created) {
+		if($created) {
+			// Que hacer cuando es creado el registro
+		} else {
+			// Que hacer cuando es actualizado el registro
+		}
+		if(!$this->updateFromfile) {
+			$this -> writeFile();
+		}
+	}
+	
+	public function afterDelete() {
+		$this -> writeFile();
 	}
 
 }
